requirements.txt
django<1.9
django-extensions
werkzeug
passlib
python-keystoneclient
python-novaclient
python-glanceclient
